pyproject.toml
[tool.poetry]
name = "f16-pagina-web"
version = "0.1.0"
description = "Página web para F16 Cargo"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
django = "^5.1.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
